Conflicts:
	Sources/Internal/Render/Highlevel/Landscape.cpp
	Sources/Internal/Render/Highlevel/Landscape.h
	Sources/Internal/Render/Material/NMaterial.cpp
	Sources/Internal/Render/Shader.cpp

Former-commit-id: 412aa6149c91f0fffcd87aa91faf2e69a9e10db7

+++ Sources/Internal/UI/UIControlBackground.h
@@ -283,7 +283,7 @@ private:
 	{
 		Vector< Vector2 > vertices;
 		Vector< Vector2 > texCoords;
-		Vector< uint32  > indeces;
+		Vector< uint16  > indeces;
 		void GenerateTileData();
 		void GenerateAxisData( float32 size, float32 spriteSize, float32 textureSize, float32 stretchCap, Vector< Vector3 > &axisData );
